LayoutTests/fast/dom/Window/Location/location-override-valueOf-using-defineGetter.html

<html>
<head>
    <script src="../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
    function getter() { throw new Error("[[Get]] should be unreachable!"); }
    shouldThrowErrorName("location.__defineGetter__('valueOf', getter)", "TypeError");
</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
